+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes for Campaign
 
+## 3.2.0 - Unreleased
+
+### Added
+
+- Added the ability to enforce spam prevention on front-end forms using Cloudflare Turnstile ([#447](https://github.com/putyourlightson/craft-campaign/issues/447)).
+- Added the `resave/campaigns`, `resave/contacts` and `resave/mailing-lists` console commands ([#481](https://github.com/putyourlightson/craft-campaign/issues/481)).
+
 ## 3.1.4 - 2024-05-06
 
 ### Fixed